Scale heap by number of locales on local node.
init_fixedHeap now scales the computed heap size by the number of locales on the local node
to avoid running out of memory. The function chpl_comm_ofi_oob_locales_on_node is called
to determine the number of locales on the local node. This function is currently only implemented
for MPI, the PMI2 and socket implementations return 1.

This resolves issue #2284.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>