Set AMO done flag via AM
On networks that do not support native atomic operations (AMO) the
Chapel runtime implements them using active messages (AM). For
none-fetching AMOs the AM handler on the target locale sets a "done"
flag on the initiating locale indicating that the AMO completed.
Previously, the handler did this via a non-blocking PUT. However, this
was found to be slower on some providers (e.g. efa) than using an AM to
set the flag. With this change the handler sends a non-blocking AM back
to the initiator to set the "done" flag.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>