Use hugepages on the hpe-cray-ex platform.
Enable use of hugepages on the hpe-cray-ex platform. The size of the heap must
be specified via CHPL_RT_HUGEPAGE_HEAP_SIZE. A hugepage heap cannot be sized
dynamically because the hugepages are faulted into memory when the heap is
allocated, preventing the runtime from striping the pages across NUMA
domains. Note that CHPL_RT_HUGEPAGE_HEAP_SIZE will be scaled appropriately
when there are co-locales.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>