Improve non-blocking AM performance in MCM mode message-order-fence
The "coforall loc in Locales do on loc" idiom is implemented by having
the initiating task serially invoke non-blocking active messages
(NBAMs) on the other locales. If the NBAM is too large to inject, the
task will block until the transmission is complete to avoid a
use-after-free error on the message.

With this change NBAMs that are too large to inject are copied by the
communication layer so that the initiating task need not wait for the
transmit to complete before sending the next message. This reduces the
time required to initiate the NBAMs and allows for more overlap in
their execution.

The functionality is controlled by three environment variables:

CHPL_RT_COMM_OFI_NBAM_THRESHOLD: NBAMs are copied if their size
(in bytes) is less than or equal to this threshold. Note that if
CHPL_RT_COMM_OFI_INJECT_AM is also set an NBAM is injected if it's
below the injection threshold. The default is 1024.

CHPL_RT_COMM_OFI_NBAM_MAX_BUFFERS: limits the number of buffers per
endpoint that are allocated to hold NBAMs, and therefore the number of
outstanding copied NBAMs per endpoint. If the number of outstanding
NBAMs on an endpoint reaches this limit subsequent sends will wait
until a buffer is available . The default is 128.

CHPL_COMM_OFI_NBAM_USE_MALLOC: if true, malloc and free are used to
allocate and deallocate the NBAM buffers. If false, the communication
layer will statically allocate CHPL_RT_COMM_OFI_NBAM_MAX_BUFFERS
buffers of size CHPL_RT_COMM_OFI_NBAM_THRESHOLD per endpoint to hold
the NBAMs. The default is true.

Implementing these features necessitated changes to the completion
counter implementation to only call fi_cntr_wait when it is ok for the
calling thread to block (i.e. fi_cntr_wait does not return until the
counter increments, which isn't always desireable).

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>