Allow per test-run customized environments
Previously, all runs of the same test (caused by having multiple
compiler and/or execution options) would use the same execution
environment, generated from the .execenv file prior to the first test
run. This made it impossible for each test run to use a different
environment. With this change, the .execenv file is read prior to each
run, and both the .preexec and .execenv files are passed the compiler
options and execution options as arguments, allowing them to specialize
the environment for the test run. This can be done by either having
the .preexec file create a new .execenv on each run, or by making
the .execenv an executable that prints the specialized environment for
the run.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>