Disable PSHM warning when running multiple processes on a host
Gasnet prints a warning to the console when running multiple processes
on a host and PSHM is not enabled. Currently, we cannot enable PSHM
because it does not support a blocking progress thread. This commit
adds a compile-time flag to disable the warning, and should be reverted
once the runtime supports a non-blocking gasnet progress thread. See
https://github.com/Cray/chapel-private/issues/3906 for details.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>