Generalize NIC selection
Generalizes NIC selection to be based on the locales' cpusets and distances
to NICs, rather than on sockets. Each locale is represented by the lowest
common ancestor of its cpuset, and each NIC is represented by its lowest
ancestor that has a cpuset. The distance between the two is the length
of the path between them in the topology. The heuristic is to find the shortest
path and assign that NIC to that locale. Then repeat with the remaining locales
and NICs until all locales have a NIC or there are no remaining NICs. If some
locales do not have a NIC the entire process is repeated until all locales
are assigned a NIC. This heuristic avoids sharing NICs unless there are
are more locales than NICs.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>