Add CHPL_FLAVOR to the Chapel environment
Add CHPL_FLAVOR to the Chapel environment, in particular the build path. This
allows having multiple versions ("flavors") of the runtime that have
differences that are not explicitely embedded in the build path, e.g.,
building a third-party library with different configurations, or building the
runtime with and without debugging symbols.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>